AdminService 통계 페이지 구현 및 TEST 완료, Socket 진행 중

+++ backend/src/main/java/com/jobPrize/dto/admin/service/PaymentCountDto.java
@@ -7,9 +7,9 @@ import lombok.Getter;
 @Getter
 public class PaymentCountDto {
 	private LocalDate date;
-	private Long payAmount;
+	private Integer payAmount;
 	
-	public PaymentCountDto(java.sql.Date date, Long payAmount) {
+	public PaymentCountDto(java.sql.Date date, Integer payAmount) {
 		this.date = date.toLocalDate();
 		this.payAmount = payAmount;
 	}
